iotsimulator/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
paho-mqtt==2.1.0
